Add toSeq in sequences to ensure immutability
In Scala 2.13 `Seq` is now an alias for `collection.immutable.Seq`. Before it was an alias for the possibly-mutable `collection.Seq`. Unfortunately some code will return collections which are `collection.Seq` and those no longer satisfy when we need the `Seq` type because they are not definitely immutable. We can get an immutable copy of the collection by doing `toSeq`.

Co-authored-by: Roberto Tyley <roberto.tyley@guardian.co.uk>