Scala 2.13: Use a stub ContentApiClient rather than a mock
This test, introduced with https://github.com/guardian/frontend/pull/25139, is intended to fail
if incompatible versions of CAPI & FAPI are used. However, it also failed when we switched
to Scala 2.13 - there seems to be some difference in the way Future.traverse behaves between
the two versions of Scala - the `null` values returned by the Mockito mock didn't cause a
problem in Scala 2.12, but threw a `NullPointerException` in Scala 2.13 (in `Future.zipwith`).

Easiest fix is to use a stub, that returns non-null objects, rather than a mock. The requests
made by this stub ContentApiClient will never succeed, but we've verified that the calls
do indeed only fail with an exception when the versions of CAPI & FAPI are incompatible.