Scala 2.13: Ensure football matches are explicitly sorted
The code previously relied on `groupBy` producing a Map that had the keys in the 'natural'
order of stage 1, stage 2 - but didn't have any sorting to enforce that. Maps are unordered
collections, and so there was never a guarantee that Scala 2.13 put Map keys in the same
order as Scala 2.12.

Co-authored-by: Roberto Tyley <roberto.tyley@guardian.co.uk>