Add test ensuring navigation.json doesn't break
For https://github.com/guardian/frontend/pull/25463, where we're adding
the Europe edition to navigation elements, we're looking to do a small
refactor of the code that generates `navigation.json` (code initially
introduced with https://github.com/guardian/frontend/pull/21409 for DCR
support - allowing DCR to know the navigation structure).

As we don't want to break the `navigation.json`, we're adding a test
that checks the data in the file has not changed, when compared against
a reference copy of the `navigation.json` commited to the repo. If
the navigation data changes (which currently only seems to happen
legitimately once or twice a year) then the reference copy needs to be
updated to make the tests pass.

The comparison is done at the json model level (`JsValue`), so simple
formatting changes will not cause the test fail.

Co-authored-by: Rowanne Kabalan <rowanne.kabalan@theguardian.com>