Remove unused Edition-based AMP mf2.json paths
https://github.com/guardian/frontend/pull/12000 in February 2016
introudced these new paths:

/container/count/:count/offset/:offset/edition/:edition/mf2.json
/container/count/:count/offset/:offset/section/:section/edition/:edition/mf2.json

However, when we check Frontend access logs, we can see no sign of these variants
(that contain `edition` in the path) being used anywhere in the last 7 days:

https://logs.gutools.co.uk/s/dotcom/goto/71a24b60-3382-11ed-b789-f1f3fdb7292e

It looks like maybe the some of the work was reverted quite soon after it
was implemented?

https://github.com/guardian/fastly-edge-cache/pull/566

As the paths aren't used, we'd like to delete them, as that means we can
delete the `getEditionFromString()` method, which has strange logic for
the edition strings and isn't used anywhere else. Otherwise, as part of
https://github.com/guardian/frontend/pull/25463 we'd have to add support
for the Europe editon to that method (and of course, that would be the
case for any new editions that came along).

Co-authored-by: Rowanne Kabalan <rowanne.kabalan@theguardian.com>