Split and map distinct switch changes in template
It seems like you can only pass a string to Play's flash message system.
But aesthetically and semantically a list of changes should be
structured as a list.
So the changes are joined into a single string to send to the template,
then split again and looped over to create separate `<li>` elements for
each.

Co-authored by: Ioanna Kokkini <ioannakok@users.noreply.github.com>