Header notifications insert and impression tracking (#25783)
Track impressions with Braze and Ophan and inserts with Ophan.

We want to send a single tracking event for each menu item and it's easier to
do this when we get all the notifications at once. So notifications are emited
in a batch when they arrive, rather than one at a time (we want a single Ophan
event for all of them).

Update braze-components to 8.1.3. This contains a fix to ensure content card
impressions are logged correctly.