Add rules to override kicker slashes
Add rules which will allow us to remove the trailing slash from kickers.
The new rules have higher specificity than the rule which adds the
slashes, because it targets elements with a combination of classes.
This allows us to add the extra class to elements where we want to
remove the slash, but keep the rest of the kicker styling.

The intention is that once this change has been finalised, the `:after`
rule will be removed from the original `fc-item__kicker` class, and then
the additional rules being added here can be removed.