Change licensing NavLink label
Update the label for the licensing link in `NavLinks` from `Guardian
content licensing site` to `Guardian licensing`.

Rebuild the `reference-navigation.json` file accordingly.

(nb. git shows a diff in the position of `tagPages`, but this is due to
how the new json is built; from looking at previous history it appears
that this key moves around in the file fairly regularly, so I don't
think it's an important diff but I chose to stick with the built
version, rather than trying to re-edit it manually and risk introducing
bugs.)