Update CricketThrottler to use OverflowStrategy.dropTail rather than OverflowStrategy.dropNew
OverflowStrategy.dropNew has been deprecated in favour of the new BoundedSourceQueue type. However, since we're creating the queue as an actorRef, we can't use this BoundedSourceQueue type.

Instead of attempting a much larger refactor to use Source.queue to create a BoundedSourceQueue, we've chosen the closest available OverflowStratergy which shouldn't be a problem as it's highly unlikely we ever exceed the buffer size.

Co-Authored-By: Daniel Clifton <110032454+DanielCliftonGuardian@users.noreply.github.com>