Refactor: make use of maybeSection more explicit
Replace mapping over `maybeSection` with a pattern match, and refactor
`redirectSection` so that it takes section as an explicit parameter,
rather than a curried, implicit one.

I believe that the `match` pattern will be more readable to future
developers coming from other languages, and is in line with how we've
been writing code elsewhere in Frontend recently.

Combined with making the `section` parameter explicit in
`redirectSection`, this should make the use of `maybeSection` clearer to
developers who are not immersed in Scala. (Using myself as an example
here!)

Co-authored-by: Jacob jacob.winch@guardian.co.uk