Rename 'guui' methods to 'DCR' methods
The `guui` naming convention no longer fits with how we
refer to the service, and is not used widely in this codebase.

I'm renaming the `getGuuiJson` and `renderGuuiJson` methods to refer to
DCR instead, to reflect current naming of the service.

I've also added comments explaining what these methods do, and for the
liveblog case I've also added a comment explaining why `lastUpdate` is
used to determine which render method to use. This is just my inference
from context though, so please take with a pinch of salt, or please feel
free to correct me!