feat: stop refreshing identity api cookies
In preperation for migrating all users to Okta we have to make sure that all users are currently using Okta sessions.

Anyone who has signed in using the sign in, reauthentication, or registration pages since
[2022-09-13](https://github.com/guardian/gateway/pull/1890) will be getting Okta sessions by default.

However there are still a small subset of users who will have stayed logged into the site since before Okta sessions were set and were getting
refreshed every 30 days.

This commit removes the ability to refresh Identity only sessions, and will now only refresh Identity cookies if an Okta session currently exists
through the use of the `refreshOktaSession` method.

This redirects to `https://profile.theguardian.com/signin/refresh`. This endpoint checks if they have a valid Okta session, if they do we refresh
the Okta session, along with the Identity cookies and redirect the user back to the page that they were on. If a user doesn't have an Okta session,
but only has existing identity cookies, then we don't refresh anything and redirect the user back. This means after a maximum of 90 days from
after this PR is merged in anyone with only an identity session will automatically be logged out.