Make request a required parameter of RequestLogger
It looks like this field was only optional to make writing tests
slightly easier, but in actual use, a request *is* always present, and
the implementation of `RequestLoggerFields` was being unnecessarily
complicated by having to deal with the `Option`.

The tests can cope with having to supply a little FakeRequest, it does
them no harm.

Co-authored-by: Jamie B <53781962+JamieB-gu@users.noreply.github.com>