Remove update-caniuse task
Running this task as part of CI directly updates the yarn.lock file if there are updates. This causes problems for the webpack bundle analyser later in compilation as the hashes on file become out of sync with those in the asset map.

I decided to move this function to a manual make command as it is not required in CI on every run and can be updated periodically.