Re-enable `TagPagesTest` (after 9 years!)
The `@DoNotDiscover` annotation (added with #6028 in September 2014)
meant that TagPagesTest was actually not getting run in CI- it was never
getting run unless it was specifically requested. It looks like this was
probably unintentional (see
https://github.com/guardian/frontend/pull/6028#issuecomment-56163278)
so I'm reinstating these tests!

I've also added a couple more tests that explicitly test the logic
for Contributor tag pages (where we prefer to index people by last name,
then first name, then just the tag's `webTitle`).