Add a request options helper
Requests to authenticated APIs need different options depending on if
the user is in the Okta experiment or not.

If the user is in the Okta experiment, the request should include an
`Authorization` header with a bearer access token value.

If the user is not in the Okta experiment, the request should include
cookies by setting the `credentials` option to `include`.

Resolves #26360.