Refactor `getUserIdentifiersFromApi` to Okta
If the user is in the Okta experiment, get the Braze UUID from the ID
Token. Otherwise, fetch from IDAPI as before.
As `brazeUuid` is the only value used, I've also renamed
the method to `getBrazeUuid` and changed the return type to `string`

Additionally, this migrates from `mergingCallback` to the Promise API

Resolves #26340