Tidy type `IdentityUser`
The `IdentityUser` type was the shape of the response from IDAPI's
`/user/me` endpoint, but the application only uses the `primaryEmailAddress`
and `statusFields.userEmailValidated` fields.

As part of the Okta migration, we want to be able to create an
`IdentityUser` with data from IDAPI or Okta's `accessToken` and
`idToken`, so we have kept just the fields being used.