Okta: Discussion API (#26403)
* Export `getAuthStatus`

* Include auth settings when `fetch`ing

Additionally, the promise should reject if the `enableDiscussionSwitch`
is off or if the user is signed out.

* Add type `UserResponse`

This type models the response shape from Discussion API's
`GET /profile/{id}` endpoint.

* Make `send`'s return type generic

Add a generic type parameter to `send` so it supports different return
types. Most functions calling send want a return type of `CommentResponse` so that
is the default, but some functions like `getUser` want a return type of `UserResponse`.

* Default to `/profile/me`

No callers of `getUser` pass an argument, so we can just use the
default.