Fix tests
It looks like the `comment-box` module used a data fetching library that
would reject the promise on some status codes. We are now using `fetch`
which only rejects the promise when a network error is encountered. A
response with a failure status code is a resolved promise.

This change updates the tests to reflect this resolving behaviour.