Fix set username when commenting (#26407)
* Refactor `updateUsername` to support Okta

If the user is signed in, make a request to IDAPI's `/user/me/username`
endpoint. Otherwise, reject the promise.

* Handle API errors

The `fail` method was never reached, because the responses in the
`fetch` API do not reject on an "unsuccessful" (non-200 response code)
request.

To fix this, we should reject promises where the `status` field is not
"ok".

* Ensure `T` is a covariant of `Response`

We want to ensure the generic parameter to `send` is one of the valid
`Response` types. Previously, it would take any type.

* Stop rejecting promises on error responses

I'd prefer to use `catch` to handle unknown errors

* Fix tests

It looks like the `comment-box` module used a data fetching library that
would reject the promise on some status codes. We are now using `fetch`
which only rejects the promise when a network error is encountered. A
response with a failure status code is a resolved promise.

This change updates the tests to reflect this resolving behaviour.