Deprecate Report Abuse Endpoints (#26454)
* Depreate routes related to reporting abusive comments

Following an investigation in #26367, we determined that these routes
and associated controllers and views can be deprecrated.

It is still possible to report abusive comments on Frontend-rendered
articles, but only if you have JS enabled. This is now parity with DCR,
which does not support no-JS discussion.

* Use a `button` to report a comment

An anchor tag is no longer semantically correct, as there is no page
navigation happening.

* Remove form method

This form should no longer make a POST request. Instead, submission is
handled in the form's event handler and a POST request is made with
`fetch`.

* Handle every submit of report abuse

I don't see why this event should only be handled once - the form might
need to be re-submitted once errors are fixed.

* Make the report abuse "reason" field mandatory

The API returns an error if you try to report a comment without
providing a reason.

* Allow signed out users to report comments

DCR does not require users to be signed in to report comments, and it
should be the same here.

The requests to other endpoints in this module should still require
authentication.

* Remove unused `DiscussionApi` methods