Send relevant `preload` and `preconnect` headers for fronts
We use the `Link` header on responses to instruct the browser to preload
assets like scripts and styles associated with the page's content.

However, all responses from `facia` currently ask browsers to preload
assets assuming that the response is HTML generated by Frontend. In
practice, this is not always the case. Facia can respond with other
types of content where we specifically _do not_ want to preload
Frontend's assets:

- JSON
- RSS
- HTML email with inlined styles and no scripts
- DCR-rendered fronts

This change refactors `FaciaController` to omit the preload header,
except for Frontend-rendered HTML fronts.

See
https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/rel/preload
for more details on asset preloading.