Update to fapi-client v4.0.6
This is a small upgrade, catching up with the recent dependency updates of
https://github.com/guardian/facia-scala-client/releases/tag/v4.0.6,
before the more extensive update in https://github.com/guardian/facia-scala-client/pull/287
is introduced.

This update has already been tested in Ophan's PromotionPoller with
https://github.com/guardian/ophan/pull/5540, successfully deployed
to Prodution.

Note that as https://github.com/guardian/frontend is _already_
using Play 2.8, it should probably be using `fapi-client-play28`,
rather than `fapi-client-play27`, so I've also updated that.