Set printable crossword cache time to 60 secs
Occasionally we needed to update the crossword data after publication. This happened recently, and while the interactive/playable crossword was very fast to update, we received some comments (https://www.theguardian.com/crosswords/cryptic/29227#comment-165327727, and others) that the `/print` version remained out of date. I was able to request a cache purge for the page to correct it in this circumstance, but it would be preferable if the cache time was sufficiently short that updates propagated quickly without intervention.