Handle JSON from SNS topic when receiving messages from the fronts tool
We are putting `FrontsUpdateSNSTopic` in front of `FrontPressToolJobQueue` (separate PR in platform will follow). Until now the queue has been receiving messages from the fronts tool directly. Once we put the topic in front of the queue, the queue will be receiving for some time messages both from the topic and and the fronts tool. With this change we're adding a way to also handle the JSON in the format the SNS topic sends it.

Co-authored-by: Georges Lebreton <georges.lebreton@guardian.co.uk>
Co-authored-by: Tom Richards <tom.richards@guardian.co.uk>
Co-authored-by: Anna Beddow <anna.beddow@gmail.com>
Co-authored-by: Alina Boghiu <alina.boghiu@gmail.com>