Rename US soccer fixtures to "schedules"
On the US soccer subnav, the word "fixtures" has been replaced with the word "schedules". The links and placement are the same.

Includes a refactor of the way this part of the nav is constructed, to avoid brittleness that might have been caused by copying and modifying the nested `children` array.