Change tested url in one of the `GalleryTemplateTest` tests
After upgrading to Play 2.9, the tests are using a new version of Selenium. With the new version of the code the outcome of `select the trail picture for the opengraph image when FacebookShareUseTrailPicFirstSwitch is ON` test affects the outcome of `select the largest main picture for the opengraph image when FacebookShareUseTrailPicFirstSwitch is OFF`. Similarly, if the tests are swapped, the second is failing. I haven't managed to figure out how to make the `TestBrowser` forget the DOM from the previous test, so for now I'm changing the tested url. I have verified by running this test with `FacebookShareUseTrailPicFirstSwitch` turned on, that the image returned is different.

See comment: https://github.com/guardian/frontend/pull/26783#issuecomment-1894346673