Modify test to check for deep equality in JSON
We were previously checking if two JSON as strings were equal but the order of the JSON fields is not guaranteed. Pulling this change out of the Play 2.9 upgrade to simplify the PR: https://github.com/guardian/frontend/pull/26783. While on Play 2.8 the test is still passing but in Play 2.9 starts failing. The order of the fields in the string is not guardanteed though, so we're now checking the actual JSON.