Remove `us-politics` from `NavLinks` list
US politics has been removed from the subnav for the duration of the 2024 US elections. The means that the subnav disappears for any pieces that are categorised as falling under this subnav item.

Removing this categorisation means that these pieces will instead be categorised another way, bringing back their subnavs.