Replace `us-politics` in `NavLinks` list
US politics has been removed from the subnav, and replaced with "US elections 2024", for the duration of the elections. This means that the subnav disappears for any pieces that are categorised as falling under this subnav item.

Replacing this categorisation with `us-elections-2024` means that these pieces will be categorised as falling under "US elections 2024" where applicable, and be categorised another way otherwise, bringing back their subnavs.