Remove Java 11 assertion from project settings
This is not necessary. We pass `-release:11` in Scala options so the Scala compiler is already going to enforce that Java 11 is the minimum supported version. Because of this PR: https://github.com/guardian/scala-steward-public-repos/pull/67 running Scala steward for frontend was failing due to this assertion because Scala Steward is running with Java 21.

Co-authored-by: Roberto Tyley <roberto.tyley@gmail.com>