Remove `PrintShopDesign`
We no longer intend to handle this as a separate `Design`. Anything that was previously `PrintShop` will now be one of the other `Design`s.

Frontend takes this property from the CAPI client and passes it through to DCAR. Therefore this change updates to a version of the CAPI client where it's been removed. The facia client has a dependency on the CAPI client, so must also be updated.

This change should be merged before removing it from DCAR. If it's removed from there while frontend is still sending it, validation errors will occur.

Preview versions for now for testing.