Move Email Newsletters To facia-rendering (#27467)
This is the only page using the old `rendering` stack. If we move it to one of the new stacks we can remove the `rendering` stack.

The four new stacks are `article-rendering`, `facia-rendering`, `tag-page-rendering` and `interactive-rendering`. This page is most similar to a front, so this change moves it to that stack. The alternative would be to create a new stack just for this page.