Substitute cricket match description for certain dismissal types
We would like to be able to use the full name of a bowler or a catcher if two players in the same team share the same surname. We also want to leave the description as it is if there is no need for subsitution, e.g. when dismissal type is "Run Out"

Co-authored-by: Roberto Tyley <52038+rtyley@users.noreply.github.com>