make the solutionAvailable field on crossword data reflect the current time (#27605)
The problem:
The crossword players offer options to help users who are struggling -
they can check whether their answer is correct, or reveal the answer if
they don't have an idea. But some crosswords offer a prize for correct
solutions, so these helpers aren't available for a period of time
(usually a week) to allow for clean(er) competition.

The CAPI data includes the solutions, regardless of whether
that period of time has passed or not. It has therefore been the
responsibility of frontend to compare the `dateSolutionAvailable` field
with the current time, and strip the solutions if required -- which it
has done for a long time without problem.

However, there is another field on the data called `solutionAvailable`
which appears to be a useful flag that can be used to decide whether to
show the check and reveal options - but this can be misleading since it
is also not updated in CAPI - so the field will show whether the
solutions should have been available _when the crossword was originally
sent to CAPI_ -- ie. for quick and cryptic crosswords it will always be
true, but for prize crosswords it will always be false (except in the
case of a very late filing!).

The solution:
It would probably be best for consumers to check that the entries do
contain solution data, or compare against the `dateSolutionAvailable`
field. That said, since this field exists and removing would be
complicated, it's probably best to make frontend also update the field
when it is building its crossword data, after comparing
`dateSolutionAvailable` to the current time, so this commit does exactly
that.