Provide DCAR JSON for galleries
Currently, when requesting JSON for a gallery, `forceDCR` is ignored. This is because the request has to be categorised as a "remote render" for DCAR JSON to be provided. However, this JSON is provided by frontend, not DCAR, so it's not accurate to call it a "remote render".

This change creates a separate case, apart from the "remote render" logic, to handle DCAR JSON for galleries.