Enable ETag-caching on Facia API requests
With https://github.com/guardian/facia-scala-client/pull/287, the Facia client
has been updated so that it can take advantage of ETag-caching, using the
https://github.com/guardian/etag-caching library already used in Frontend since
August 2023 with https://github.com/guardian/frontend/pull/26338.

As some changes were necessary in the `etag-caching` library to facilitate this,
we need to update the version of `etag-caching` used by Frontend to v7.0.0.

We're also upgrading the S3 client (used by Frontend's instance of Facia's
`ApiClient`) to AWS SDK v2, as `etag-caching` already provides a convenient
implementation of `S3ByteArrayFetching` (see https://github.com/guardian/etag-caching/pull/65)
for AWS SDK v2 - if we continued using AWS SDK v1 here in Frontend, we'd have
to create an implementation for it.

## See also

* https://github.com/guardian/maintaining-scala-projects/issues/9
* https://github.com/guardian/facia-scala-client/pull/287