Update Facia and ETag-caching library
With https://github.com/guardian/facia-scala-client/pull/287, the Facia client has
been updated so that it can take advantage of ETag-caching, using the
https://github.com/guardian/etag-caching library already used in Frontend since
August 2023 with https://github.com/guardian/frontend/pull/26338.

As some changes were necessary to the `etag-caching` library to facilitate
this, we need to update the version of `etag-caching` used by Frontend to v7.0.0.

Although this change upgrades to the latest version of the Facia client, it doesn't
take advantage of the new ETag-caching support there - that can come in a later PR,
as it does involve switching to AWS SDK v2 (not everywhere in Frontend, just for
where we use the Facia client, but it's still a significant change as AWS STS
credential creation has got more elaborate with AWS SDK v2).