Remove Play plugin from root project
The root project is used to aggregate the other projects for compilation
and testing. It's not run as an application in its own right, and
therefore does not need to include the Play plugin directly. Removing it
also removes the need for the root project to have its own dependency
tree.

Co-authored-by: Roberto Tyley <52038+rtyley@users.noreply.github.com>