Split `DotcomRenderingFootballDataModel`
Split this generic `case class` into a `trait` and two `case class`es,
one each for the football match list and football tables. A generic
suggests `DotcomRenderingFootballDataModel` can be a container for
anything, whereas in reality there are only two models. This also makes
it easier to differentiate the two models (e.g. in pattern matches),
which is not possible with a generic.

As part of this, moved some of the JSON encoding implicits into the two
new companion objects, as not all of these are required for both.

Also removed a duplicate edition lookup. We were calculating the edition
twice, using two different APIs (the `apply` and `edition` methods).
This consolidates on the `apply` method.

Co-authored-by: Marjan Kalanaki <marjan.kalanaki@guardian.co.uk>