Update Netty To `4.1.118.Final` (#27834)
* Update netty to `4.1.118.Final`

Netty is a transitive dependency of the AWS SDK, so this has been
updated to the latest version. It's also a transitive dependency of
Play, but the latest version of Play still depends on `4.1.115.Final`,
so this change adds direct dependencies on netty to force an eviction.

* Removed the netty overrides

Merging the latest `main` has brought in Play 3.0.7, which depends on
netty 4.1.119.Final. With both the update to the AWS SDK and Play we no
longer need to evict older versions.

Co-authored-by: Daniel Clifton <110032454+DanielCliftonGuardian@users.noreply.github.com>
Co-authored-by: Marjan Kalanaki <marjan.kalanaki@guardian.co.uk>
Co-authored-by: James Mockett <1166188+jamesmockett@users.noreply.github.com>