Simplified `LeagueTableController`
It's not yet necessary to introduce the `remoteRenderer`, as we're just
returning the DCAR JSON in this change. In addition, using
`RemoteRender` in this way would remove the ability to get non-DCAR JSON
from this endpoint once DCAR support for these pages is switched on. We
should instead only return the DCAR JSON when `forceDCR` is set.

Also removed other features like `Action.async` and `Future`, as they're
not needed unless we're calling DCAR.