comment out the test calling the more endpoints
The more endpoints are temporarily removed.
These tests are calling the removed endpoints when the more button is
clicked. The more button is supposed to call the json routes but this
call is controlled in the javascript. It seems the test doesn't have
access to the js and clicking the more button is using the href link
which is the url without '.json' at the end of it.