Apply default boost levels where available
This will apply the correct boost level to backfilled content in flexible general containers. It will also ensure that the limits of each group in a flexible general container are respected.

As much of the required logic as is possible is delegated to the facia-scala-client to keep MAPI and Frontend in sync