Migrate to v2 of the AWS Java SDK for S3
The first part of the S3 migration.

Panda settings require an instance of an `S3BucketLoader`. Panda comes
with a default implementation for v1, but not v2, so this change creates
one.

This first requires a synchronous S3 client, which has been added to the
`AWSv2` object alongside the existing asynchronous S3 client. Creating
it here, rather than in the controller as before, allows it to be used
more widely.

Then the `getObject` method can be applied, with a `ResponseTransformer`
to retrieve the "content" of the response, i.e. the object itself.

We're removing `S3Client` as it became unused with
https://github.com/guardian/frontend/pull/27391

We've also added support for `getPresignedUrl`, which was introduced with
https://github.com/guardian/frontend/pull/28000 after we started this PR!

Co-authored-by: Roberto Tyley <52038+rtyley@users.noreply.github.com>