Make sure internal redirects for content redirect to the correct URL
Since the evolving URLs feature (where a piece of content's URL may
change over time), it is possible to create a redirect loop.

A piece can be launched with a "normal" URL, (which looks like
`/section/year/month/day/slug`), then evolved to an odder (but still
very much legitimate) URL, like `/slug`. This looks indistinguishable
from a fronts URL, so requests will be routed initially to facia, but
once it is discovered that it is _not_ a front but instead a piece of
content, it will be internally redirected. Currently this internal
redirect sends the request by using the content ID, but crucially the
content ID is not updated with the evolving URLs feature. So when a
request comes through with the content's ID, it is considered to be a
request for the outdated URL, and an external redirect is issued,
sending the requester to the new URL.

But this new URL is the exact same that the initial request was for,
meaning that for the reader's browser a request for `/slug` redirects to
`/slug`, meaning we immediately enter an indefinite redirect loop.

If we instead perform the internal redirect using the `webUrl` property,
we should redirect to the article service using the latest current URL
and avoid any need for further redirections.