Make Archive's gallery redirects temporary
Like all content, if someone accidentally attempts to open a Gallery
before it's been published, we'll get a 404 from CAPI in the usual
rendering path, which will internally redirect the request to Archive in
case the content is in the archive instead. However Archive rewrites
requests for `.../gallery/...` to `.../pictures/...` using a *permanent*
redirect. The redirected request will 404 since the content's not in the
archive, which is expected, but since the redirect was permanent, the
reader will now always be sent to this `pictures` URL, *even when the
content has been published* on its intended `/gallery/` URL, and it
seems almost impossible to remove a permanent redirect once it's been
established in a browser.

Instead, use a temporary redirect. This might incur some slightly
increased requests for archived galleries which would otherwise have
been redirected by browsers but that feels pretty uncommon, and is a
worthwhile tradeoff for preventing readers accidentally locking
themselves out of content.