Log "info" when crossword not found
A request for a crossword that doesn't exist in CAPI results in a 404.
This is currently being logged as an error, but that suggests it's a
problem with the platform that we should fix. In reality, a piece of
content that does not exist is not something that's likely to change,
or that we can "fix", so it should be logged as a piece of information
rather than an error.