Prevent error on a sections request
Summary

`/sections` triggers a CAPI sections query, but frontend tries to
parse the response as an `ItemResponse` and crashes. Since `/sections`
is not a page that exists on dotcom, we can avoid this problem by
responding with a 404 (not found).

Details

Requests to `/sections` on dotcom currently result in a 500 error. This
happens because this path is interpreted as an index page by the
applications app's routes file, which results in a call to the
`IndexController.render` method. This is a concrete method on the
`IndexControllerCommon` trait, which `IndexController` extends. `render`
in turn calls the `renderItem` method, also implemented on
`IndexControllerCommon`.

For this path, the `renderItem` method currently attempts to make an
`ItemRequest` to CAPI, and then calls the `renderFaciaFront` method with
the result. This method is defined on the extending `IndexController`
from the applications app, and ultimately makes a request to DCAR to
render a tag page.

However, `/sections` is not a tag page. It doesn't have a particular
meaning on dotcom, but it does have a special meaning to CAPI, where it
returns a list of sections. There's a separate `SectionsQuery` for this,
but in this case frontend is sending CAPI an `ItemQuery`. However,
because this is implemented as an HTTP GET request with `/sections` as
the path, CAPI believes it is being queried for sections, and returns
a `SectionsResponse`. Frontend is expecting an `ItemResponse` and
therefore the Thrift deserialisation code crashes because the fields do
not match.

The solution is to explicitly respond to requests for `/sections` with a
404 (not found), before we get to the point where a CAPI request is
made.